Ensure a clip connect(false) is sent
This event was sometimes not send when the mouse was held down and then
moved, in this case a "drag" event would occur and no mouseup was sent,
so the clip was never released. We now ignore drag events completely
(avoiding the ugly drag image) and release when the mouse leaves the
area.